refactor: move auth config to config package
